move LadderMode to mod.rs
