+++ src/fh_va/ladder.rs
@@ -10,6 +10,8 @@ use crate::fh_va::FilterParams;
 use std::simd::*;
 use std::sync::Arc;
 
+use super::{LadderMode, get_ladder_mix};
+
 #[allow(dead_code)]
 #[derive(PartialEq, Clone, Copy)]
 enum EstimateSource {
@@ -262,78 +264,3 @@ impl LadderFilter {
         sum
     }
 }
-
-#[derive(Debug, PartialEq)]
-pub enum LadderMode {
-    Lp6,
-    Lp12,
-    Lp18,
-    Lp24,
-    Hp6,
-    Hp12,
-    Hp18,
-    Hp24,
-    Bp12,
-    Bp24,
-    N12,
-}
-impl std::fmt::Display for LadderMode {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            LadderMode::Lp6 => write!(f, "Lp6"),
-            LadderMode::Lp12 => write!(f, "Lp12"),
-            LadderMode::Lp18 => write!(f, "Lp18"),
-            LadderMode::Lp24 => write!(f, "Lp24"),
-            LadderMode::Hp6 => write!(f, "Hp6"),
-            LadderMode::Hp12 => write!(f, "Hp12"),
-            LadderMode::Hp18 => write!(f, "Hp18"),
-            LadderMode::Hp24 => write!(f, "Hp24"),
-            LadderMode::Bp12 => write!(f, "Bp12"),
-            LadderMode::Bp24 => write!(f, "Bp24"),
-            LadderMode::N12 => write!(f, "N12"),
-        }
-    }
-}
-pub fn get_ladder_mix(mode: LadderMode) -> [f32; 5] {
-    let mix;
-    match mode {
-        LadderMode::Lp6 => {
-            mix = [0., -1., 0., -0., 0.];
-        }
-        LadderMode::Lp12 => {
-            mix = [0., -0., 1., -0., 0.];
-        }
-        LadderMode::Lp18 => {
-            mix = [0., -0., 0., -1., 0.];
-        }
-        LadderMode::Lp24 => {
-            mix = [0., -0., 0., -0., 1.];
-        }
-        LadderMode::Hp6 => {
-            mix = [1., -1., 0., -0., 0.];
-        }
-        LadderMode::Hp12 => {
-            mix = [1., -2., 1., -0., 0.];
-        }
-        LadderMode::Hp18 => {
-            mix = [1., -3., 3., -1., 0.];
-        }
-        LadderMode::Hp24 => {
-            mix = [1., -4., 6., -4., 1.];
-        }
-        LadderMode::Bp12 => {
-            mix = [0., -1., 1., -0., 0.];
-        }
-        LadderMode::Bp24 => {
-            mix = [0., -0., 1., -2., 1.];
-        } 
-        LadderMode::N12 => {
-            mix = [1., -2., 2., -0., 0.];
-
-        },
-          
-    }
-    mix
-}
-
-
